made classes with common functionality shared between XDG shell versions
