+++ src/server/frontend_wayland/CMakeLists.txt
@@ -17,7 +17,7 @@ set(
   wl_keyboard.cpp               wl_keyboard.h
   wl_pointer.cpp                wl_pointer.h
   wl_touch.cpp                  wl_touch.h
-  xdg_shell_v6.cpp              xdg_shell_v6.h
+  xdg_shell_base.cpp            xdg_shell_base.h
 )
 
 add_library(
